refactor(examples): タグ・投稿ユーティリティを lib/ に抽出して関心を分離
getAllTags を astro-blog / astro-blog-simple の lib/tags.ts に、
parsePostId を astro-blog の lib/posts.ts に移動し、
config/ はデータ定義のみに絞った。
