+++ examples/astro-blog/src/components/TagCloud.astro
@@ -2,13 +2,10 @@
 import { getCollection } from 'astro:content';
 import { Cluster } from 'lism-css/astro';
 import Tag from '@/components/Tag.astro';
+import { getAllTags } from '@/lib/tags';
 
 const posts = await getCollection('posts');
-const tagSet = new Set<string>();
-for (const post of posts) {
-  for (const tag of post.data.tags) tagSet.add(tag);
-}
-const tags = Array.from(tagSet).sort();
+const tags = getAllTags(posts);
 
 const { tagBgc = 'base-2' } = Astro.props;
 ---
