+++ examples/astro-blog/src/config/categories.ts
@@ -14,7 +14,7 @@ export const CATEGORIES: Record<CategoryKey, Category> = {
   dev: {
     key: 'dev',
     label: 'DEV',
-    description: 'Lism CSS や Astro まわりの開発メモ',
+    description: '日々の開発メモ',
   },
   life: {
     key: 'life',
@@ -24,18 +24,3 @@ export const CATEGORIES: Record<CategoryKey, Category> = {
 };
 
 export const CATEGORY_KEYS: CategoryKey[] = ['dev', 'life'];
-
-export function isCategoryKey(value: string): value is CategoryKey {
-  return value === 'dev' || value === 'life';
-}
-
-/**
- * post.id（例: "dev/foo"）から category と slug を分解する。
- */
-export function parsePostId(id: string): { category: CategoryKey; slug: string } {
-  const [category, ...rest] = id.split('/');
-  if (!isCategoryKey(category)) {
-    throw new Error(`Unknown category in post id: ${id}`);
-  }
-  return { category, slug: rest.join('/') };
-}
