+++ examples/astro-blog/src/pages/tag/[tag]/[...page].astro
@@ -6,15 +6,12 @@ import ArchiveHeader from '@/components/ArchiveHeader.astro';
 import PostList from '@/components/PostList.astro';
 import Pager from '@/components/Pager.astro';
 import { siteConfig } from '@/config/site';
+import { getAllTags } from '@/lib/tags';
 
 export const getStaticPaths = (async ({ paginate }) => {
   const allPosts = await getCollection('posts');
-  const tagSet = new Set<string>();
-  for (const post of allPosts) {
-    for (const tag of post.data.tags) tagSet.add(tag);
-  }
 
-  return Array.from(tagSet).flatMap((tag) => {
+  return getAllTags(allPosts).flatMap((tag) => {
     const posts = allPosts.filter((p) => p.data.tags.includes(tag)).sort((a, b) => b.data.date.localeCompare(a.data.date));
     return paginate(posts, {
       params: { tag },
